# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Missing Semester Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  learner: 'learner',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface Lesson {
  id: number;
  moduleNum: number;
  lessonNum: number;
  title: string;
  description: string;
  requiresApiKeys: boolean;
}

export interface Module {
  moduleNum: number;
  title: string;
  description: string;
  lessons: Lesson[];
}

export interface LessonDetail {
  id: number;
  moduleNum: number;
  lessonNum: number;
  title: string;
  description: string;
  objectives: string[];
  notebookPath: string;
  requiresApiKeys: boolean;
  requiredServices: string[];
}

export interface Progress {
  id: number;
  userId: number;
  lessonId: number;
  cellsRun: number;
  totalCells: number;
  /** @nullable */
  completedAt?: string | null;
  lastActiveAt: string;
}

export interface ProgressUpdate {
  cellsRun?: number;
  totalCells?: number;
  completed?: boolean;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  starting: 'starting',
  running: 'running',
  paused: 'paused',
  stopped: 'stopped',
  error: 'error',
} as const;

export interface Session {
  id: number;
  userId: number;
  lessonId: number;
  status: SessionStatus;
  /** @nullable */
  containerUrl?: string | null;
  /** @nullable */
  ecsTaskArn?: string | null;
  createdAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export interface SessionOrNull {
  session: Session | null;
}

export interface SessionInput {
  lessonId: number;
}

export interface ApiKey {
  id: number;
  userId: number;
  service: string;
  maskedKey: string;
  createdAt: string;
}

export interface ApiKeyInput {
  /** @minLength 1 */
  key: string;
}

export interface AdminStats {
  totalLearners: number;
  sessionsToday: number;
  sessionsThisWeek: number;
  lessonsCompleted: number;
  activeSessions: number;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  role: string;
  createdAt: string;
  /** @nullable */
  lastActiveAt?: string | null;
  lessonsCompleted: number;
  lessonsStarted: number;
}

export interface LessonFunnelItem {
  lessonId: number;
  lessonTitle: string;
  moduleNum: number;
  lessonNum: number;
  started: number;
  completed: number;
}

export interface AdminSession {
  id: number;
  userId: number;
  userEmail: string;
  userName: string;
  lessonId: number;
  lessonTitle: string;
  status: string;
  createdAt: string;
  /** @nullable */
  endedAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
}