# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Missing Semester Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSession {
  id: number;
  userId: number;
  userEmail: string;
  userName: string;
  lessonId: number;
  lessonTitle: string;
  status: string;
  createdAt: Date;
  /** @nullable */
  endedAt?: Date | null;
  /** @nullable */
  durationMinutes?: number | null;
}