# lib/api-zod/src/generated/types/module.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Missing Semester Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from './lesson';

export interface Module {
  moduleNum: number;
  title: string;
  description: string;
  lessons: Lesson[];
}