# lib/api-zod/src/generated/types/progress.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Missing Semester Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Progress {
  id: number;
  userId: number;
  lessonId: number;
  cellsRun: number;
  totalCells: number;
  /** @nullable */
  completedAt?: Date | null;
  lastActiveAt: Date;
}