# lib/api-zod/src/generated/types/progressUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Missing Semester Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressUpdate {
  cellsRun?: number;
  totalCells?: number;
  completed?: boolean;
}